optimise CRUD customer

+++ src/Controller/CustomerController.php
@@ -8,11 +8,14 @@ use App\Entity\Customer;
 use App\Entity\User;
 use App\Repository\CustomerRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Lcobucci\JWT\Token;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 
 
@@ -23,16 +26,24 @@ use Symfony\Component\Serializer\SerializerInterface;
  */
 class CustomerController extends AbstractController
 {
+
     /**
      * @Route(name="api_customers_collection_get", methods={"GET"})
      * @param CustomerRepository $customerRepository
      * @param SerializerInterface $serializer
+     * @param TokenStorageInterface $tokenStorage
      * @return JsonResponse
      */
-    public function collection(CustomerRepository $customerRepository, SerializerInterface $serializer): JsonResponse
+    public function collection(
+        CustomerRepository $customerRepository,
+        SerializerInterface $serializer,
+        TokenStorageInterface $tokenStorage
+    ): JsonResponse
     {
+        /** @var User $user */
+        $user = $tokenStorage->getToken()->getUser();
         return new JsonResponse(
-            $serializer->serialize($customerRepository->findAll(), "json", ["groups" => "get"]),
+            $serializer->serialize($customerRepository->findBy(["user" => $user->getId()]), "json", ["groups" => "get"]),
             JsonResponse::HTTP_OK,
             [],
             true
@@ -44,18 +55,28 @@ class CustomerController extends AbstractController
      * @Route("/{id}", name="api_customers_item_get", methods={"GET"})
      * @param Customer $customer
      * @param SerializerInterface $serializer
+     * @param TokenStorageInterface $tokenStorage
      * @return JsonResponse
      */
-    public function item(Customer $customer, SerializerInterface $serializer): JsonResponse
+    public function item(Customer $customer, SerializerInterface $serializer, TokenStorageInterface $tokenStorage): JsonResponse
     {
-        return new JsonResponse(
-            $serializer->serialize($customer, "json", ["groups" => "get"]),
-            JsonResponse::HTTP_OK,
-            [],
-            true
-        );
+        if ($customer->getUser() === $tokenStorage->getToken()->getUser())
+        {
+            return new JsonResponse(
+                $serializer->serialize($customer, "json", ["groups" => "get"]),
+                JsonResponse::HTTP_OK,
+                [],
+                true
+            );
+        }else{
+            return new JsonResponse(
+                ['code' => JsonResponse::HTTP_FORBIDDEN, 'status' => "You do not have permission for this user"],
+                JsonResponse::HTTP_FORBIDDEN
+            );
+        }
 
     }
+
 
     /**
      * @Route(name="api_customers_collection_post", methods={"POST"})
@@ -63,20 +84,22 @@ class CustomerController extends AbstractController
      * @param SerializerInterface $serializer
      * @param EntityManagerInterface $entityManager
      * @param UrlGeneratorInterface $urlGenerator
+     * @param TokenStorageInterface $tokenStorage
      * @return JsonResponse
      */
     public function post(
         Request $request,
         SerializerInterface $serializer,
         EntityManagerInterface $entityManager,
-        UrlGeneratorInterface $urlGenerator
+        UrlGeneratorInterface $urlGenerator,
+        TokenStorageInterface $tokenStorage
     ): JsonResponse
     {
         /** @var Customer $customer */
         $customer = $serializer->deserialize($request->getContent(), Customer::class, 'json');
 
         /** @var User $user */
-        $user = $entityManager->getRepository(User::class)->findOneBy(["id" => 1]);
+        $user = $tokenStorage->getToken()->getUser();
 
         $customer->setUser($user);
 
@@ -96,15 +119,27 @@ class CustomerController extends AbstractController
      * @Route("/{id}", name="api_customers_item_delete", methods={"DELETE"})
      * @param Customer $customer
      * @param EntityManagerInterface $entityManager
+     * @param TokenStorageInterface $tokenStorage
      * @return JsonResponse
      */
-    public function delete(Customer $customer, EntityManagerInterface $entityManager): JsonResponse
+    public function delete(Customer $customer, EntityManagerInterface $entityManager, TokenStorageInterface $tokenStorage): JsonResponse
     {
-        $entityManager->remove($customer);
-        $entityManager->flush();
+        if ($customer->getUser() === $tokenStorage->getToken()->getUser())
+        {
+            $entityManager->remove($customer);
+            $entityManager->flush();
 
-        return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+            return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+        }else{
+            return new JsonResponse(
+                ['code' => JsonResponse::HTTP_FORBIDDEN, 'status' => "You do not have permission for this user"],
+                JsonResponse::HTTP_FORBIDDEN
+            );
+        }
+
 
     }
+
+
 
 }
